Move parsing helpers from service to new PublicationFieldValue class
